+++ pallets/subtensor/src/weights.rs
@@ -81,7 +81,7 @@ impl<T: Config> Pallet<T> {
         netuid: u16,
         uids: Vec<u16>,
         values: Vec<u16>,
-        salt: Vec<u8>,
+        salt: Vec<u16>,
         version_key: u64,
     ) -> DispatchResult {
         let who = ensure_signed(origin.clone())?;
